+++ src/main/java/ch/uzh/ifi/hase/soprafs23/model/Notifier.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 public interface Notifier {
 
-    public void publishRoundScores(Long user_id,Map<String, ArrayList<Map<String, Integer>>> roundScores);
+    public void publishRoundScores(Long user_id,Map<String, ArrayList<Map<String, Double>>> roundScores);
 
     public void publishGameScores(Map<String, Integer> placement);
     public void publishFinalScores(Map<String, Integer> placement);
